update all article/themes page
